+++ package.json
@@ -16,7 +16,8 @@
     "openai-edge": "^1.2.2",
     "prettier": "^3.1.0",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "react-icons": "^5.0.1"
   },
   "devDependencies": {
     "@types/node": "^20",
